Move double register test into poll's test module
Reduces the output size.
